Initial support for running NetBox in Qemu with 'make run'

+++ board/common/post-image.sh
@@ -7,8 +7,41 @@
 imagesh=$BR2_EXTERNAL_NETBOX_PATH/utils/image.sh
 fitimagesh=$BR2_EXTERNAL_NETBOX_PATH/utils/fitimage.sh
 
+ext2=$BINARIES_DIR/rootfs.ext2
 squash=$BINARIES_DIR/rootfs.squashfs
 img=$BINARIES_DIR/$BR2_EXTERNAL_ID-$NETBOX_TYPE-${NETBOX_PLAT}
+
+qemucfg="${BINARIES_DIR}/qemu.cfg"
+
+QEMU_ARCH=""
+case $BR2_ARCH in
+    powerpc)
+	#QEMU_ARCH=ppc
+	;;
+    arm)
+	#QEMU_ARCH=arm
+	;;
+    aarch64)
+	#QEMU_ARCH=arm
+	;;
+    x86_64)
+	QEMU_ARCH=i386
+	QEMU_NIC=rtl8139
+	QEMU_SCSI=virtio-scsi-pci
+	QEMU_MACH="q35,accel=kvm -smp 2 -watchdog i6300esb -cpu host -enable-kvm -rtc clock=host"
+	;;
+    *)
+	;;
+esac
+
+cat <<EOF > $qemucfg
+QEMU_ARCH=$QEMU_ARCH
+QEMU_MACH="$QEMU_MACH"
+QEMU_NIC=$QEMU_NIC
+QEMU_SCSI=$QEMU_SCSI
+
+QEMU_KERNEL=${BINARIES_DIR}/$(basename $BINARIES_DIR/*Image)
+EOF
 
 err=0
 if [ -n "$RELEASE" ]; then
@@ -24,16 +57,24 @@ if [ -n "$RELEASE" ]; then
     fi
 fi
 
-$imagesh $squash $img.img
+if [ "$BR2_TARGET_ROOTFS_SQUASHFS" = "y" ]; then
+    $imagesh $squash $img.img
 
-if [ "$NETBOX_IMAGE_FIT" ]; then
-    $fitimagesh $NETBOX_PLAT $squash $img.itb
+    if [ "$NETBOX_IMAGE_FIT" ]; then
+	$fitimagesh $NETBOX_PLAT $squash $img.itb
+    fi
+
+    echo "QEMU_INITRD=${BINARIES_DIR}/rootfs.squashfs" >>$qemucfg
+fi
+
+if [ "$BR2_TARGET_ROOTFS_EXT2" = "y" ]; then
+    echo "QEMU_DISK=${BINARIES_DIR}/rootfs.ext2" >>$qemucfg
 fi
 
 # Set TFTPDIR, in your .bashrc, or similar, to copy the resulting image
 # to your FTP/TFTP server directory.  Notice the use of scp, so you can
 # copy the image to another system.
-if [ -n "$TFTPDIR" ]; then
+if [ -n "$TFTPDIR" -a -e $img.img ]; then
     echo "xfering '$img' -> '$TFTPDIR/$fn'"
     scp -B $img.img $TFTPDIR
 fi
